merge master

+++ tests/test_rotated.py
@@ -14,9 +14,9 @@ import unittest
 import numpy as np
 import scipy.ndimage
 from parameterized import parameterized
-from tests.utils import NumpyImageTestCase2D, NumpyImageTestCase3D
 
 from monai.transforms import Rotated
+from tests.utils import NumpyImageTestCase2D, NumpyImageTestCase3D
 
 TEST_CASES_2D = [
     (-30, False, "bilinear", "border", False),
@@ -78,7 +78,7 @@ class TestRotated3D(NumpyImageTestCase3D):
         expected = scipy.ndimage.rotate(
             self.imt[0, 0], angle, (0, 2), not keep_size, order=_order, mode=_mode, prefilter=False
         )
-        np.testing.assert_allclose(expected, rotated["img"][0], atol=1e-3)
+        np.testing.assert_allclose(expected.astype(np.float32), rotated["img"][0], atol=1e-3)
 
         expected = scipy.ndimage.rotate(
             self.segn[0, 0], angle, (0, 2), not keep_size, order=0, mode=_mode, prefilter=False
